pyproject.toml
[project]
name = "ruuvi-mcp"
version = "0.1.0"
description = "MCP server for Ruuvi sensor data"
authors = [
    {name = "Juha Pellotsalo"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27.0",
    "mcp>=1.25.0",
    "paho-mqtt>=2.0.0",
    "pyyaml>=6.0.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[project.scripts]
ruuvi-cli = "cli.app:main"
ruuvi-mcp = "ruuvi_mcp.server:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["ruuvi_mcp*", "ruuvi_data*", "cli*"]